Rework to allow for multiple notes, almost done

+++ node-express/index.js
@@ -15,9 +15,9 @@ app.use(bodyParser.json());
 app.post('/:username', (req, res) => {
     let username = req.params.username;
     let filePath = "./Notes/" + username + ".txt";
-    let note = req.body.note;
+    let note = "&&" + req.body.note;        // && Will be used as the delimiter for each of our notes so we can split the string 
 
-    fs.writeFile(filePath, note, err => {
+    fs.appendFile(filePath, note, err => {
         if (err) {
             console.error(err);
             res.status(404).send('Unable to add note');
@@ -73,19 +73,53 @@ app.put('/:username', (req, res) => {
 })
 
 // Delete note with DELETE request
-app.delete('/:username', (req, res) => {
+app.delete('/:username/:ids', (req, res) => {
     let username = req.params.username;
+    let ids = req.params.ids;
+    let notes;
+    let updatedNotes = '';
 
     let filePath = "./Notes/" + username + ".txt";      //Creates the right file path for the notes in our local Notes folder
 
-    fs.unlink(filePath, (err) => {
+    fs.readFile(filePath, 'utf8', (err, message) => {
         if (err) {
             console.error(err);
-            res.status(404).send('No note to remove');
-            return;
         }
-        else
-        res.status(200).send('Note successfully deleted');
+        else {
+            notes = message.split('&&');
+            console.log(notes);
+            console.log(ids);
+
+            for (i = 1; i < notes.length; i++){
+                if (ids.includes(i))        // Skips over the indices that were selected to be removed
+                    continue;
+                else    
+                    updatedNotes += ('&&' + notes[i]);
+            }
+
+            if (updatedNotes == ''){
+                fs.unlink(filePath, (err) => {      // Deletes .txt file if all notes are chosen for deletion
+                    if (err) {
+                        console.error(err);
+                        res.status(404).send('No notes to remove');
+                        return;
+                    }
+                    else
+                        res.status(200).send('Notes successfully deleted');
+                });
+            }
+            else {
+                fs.writeFile(filePath, updatedNotes, err => {       // Deletes notes by writing all notes not selected for deletion to .txt file
+                    if (err) {
+                        console.error(err);
+                        res.status(404).send('Unable to delete notes');
+                        return;
+                    }
+                    else
+                        res.status(200).send('Notes successfully deleted');
+                });
+            }
+        }
     });
 
 })
